+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * [Improved] Add `highlight` and `fuzzy` options to ext-filter
   * [Improved] Add `fireActivate` option to ext-persist (default: true)
   * [Fixed] #475 Font color while editing node title with bootstrap skin
+  * [Fixed] #484 Glyph plugin: Missing margin-left for span.fancytree-custom-icon
 
 # 2.11.0 / 2015-07-26
   * [Changed] Adding `fancytree-plain` class to container (if not table), allowing for more efficient css